pyproject.toml
[project]
name = "fastapi-boilerplate"
version = "0.1.0"
description = "FastAPI boilerplate with auth, settings, logging, and Docker"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "faker>=37.5.3",
    "fastapi[all]>=0.115.12",
    "loguru>=0.7.3",
    "psycopg2>=2.9.10",
    "psycopg[binary]>=3.2.9",
    "python-jose>=3.5.0",
    "ruff>=0.12.8",
    "sqlmodel>=0.0.24",
    "starlette>=0.47.2",
    "uvicorn>=0.34.2",
]

[tool.ruff]
# Same as Black.
line-length = 88

# Assume Python 3.12
target-version = "py312"

[tool.ruff.lint]
# Enable all rules by default
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = ["B008"]

# Allow autofix for all enabled rules (when `--fix` is used)
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.lint.isort]
known-first-party = ["fastapi_boilerplate"]

[tool.ruff.format]
# Use double quotes for strings.
quote-style = "double"

# Indent with spaces, rather than tabs.
indent-style = "space"

# Respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, respect magic trailing commas.
line-ending = "auto"

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
